Split templates by event type
