+++ src/constants/MoveHistory/EventTemplates/CreateMTOShipment/createMTOShipment.test.jsx
@@ -2,7 +2,7 @@ import getTemplate from 'constants/MoveHistory/TemplateManager';
 import a from 'constants/MoveHistory/Database/Actions';
 import o from 'constants/MoveHistory/UIDisplay/Operations';
 import t from 'constants/MoveHistory/Database/Tables';
-import e from 'constants/MoveHistory/EventTemplates/MTOShipment/createMTOShipment';
+import e from 'constants/MoveHistory/EventTemplates/CreateMTOShipment/createMTOShipment';
 
 describe('when given a Create mto shipment history record', () => {
   const item = {
